# Set parents during parse, not bind