--- a/internal/execute/tsc.go
+++ b/internal/execute/tsc.go
@@ -78,7 +78,6 @@ func fmtMain(sys System, input, output string) ExitStatus {
 		Path:             pathified,
 		JSDocParsingMode: ast.JSDocParsingModeParseAll,
 	}, text, core.GetScriptKindFromFileName(string(pathified)))
-	ast.SetParentInChildren(sourceFile.AsNode())
 	edits := format.FormatDocument(ctx, sourceFile)
 	newText := applyBulkEdits(text, edits)