--- a/internal/printer/printer_test.go
+++ b/internal/printer/printer_test.go
@@ -614,7 +614,6 @@ func TestParenthesizeDecorator(t *testing.T) {
 		},
 	))
 
-	ast.SetParentInChildren(file)
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "@(a + b)\nclass C {\n}")
 }
@@ -651,7 +650,7 @@ func TestParenthesizeComputedPropertyName(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "class C {\n    [(a, b)];\n}")
 }
@@ -681,7 +680,7 @@ func TestParenthesizeArrayLiteral(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "[(a, b)];")
 }
@@ -709,7 +708,7 @@ func TestParenthesizePropertyAccess1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a, b).c;")
 }
@@ -736,7 +735,7 @@ func TestParenthesizePropertyAccess2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a?.b).c;")
 }
@@ -762,7 +761,7 @@ func TestParenthesizePropertyAccess3(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(new a).b;")
 }
@@ -790,7 +789,7 @@ func TestParenthesizeElementAccess1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a, b)[c];")
 }
@@ -817,7 +816,7 @@ func TestParenthesizeElementAccess2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a?.b)[c];")
 }
@@ -843,7 +842,7 @@ func TestParenthesizeElementAccess3(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(new a)[b];")
 }
@@ -872,7 +871,7 @@ func TestParenthesizeCall1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a, b)();")
 }
@@ -900,7 +899,7 @@ func TestParenthesizeCall2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a?.b)();")
 }
@@ -927,7 +926,7 @@ func TestParenthesizeCall3(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(new C)();")
 }
@@ -957,7 +956,7 @@ func TestParenthesizeCall4(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "a((b, c));")
 }
@@ -984,7 +983,7 @@ func TestParenthesizeNew1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "new (a, b)();")
 }
@@ -1011,7 +1010,7 @@ func TestParenthesizeNew2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "new (C());")
 }
@@ -1039,7 +1038,7 @@ func TestParenthesizeNew3(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "new C((a, b));")
 }
@@ -1068,7 +1067,7 @@ func TestParenthesizeTaggedTemplate1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a, b) ``;")
 }
@@ -1096,7 +1095,7 @@ func TestParenthesizeTaggedTemplate2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a?.b) ``;")
 }
@@ -1125,7 +1124,7 @@ func TestParenthesizeTypeAssertion1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "<T>(a + b);")
 }
@@ -1152,7 +1151,7 @@ func TestParenthesizeArrowFunction1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "() => ({});")
 }
@@ -1184,7 +1183,7 @@ func TestParenthesizeArrowFunction2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "() => ({}.a);")
 }
@@ -1209,7 +1208,7 @@ func TestParenthesizeDelete(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "delete (a + b);")
 }
@@ -1234,7 +1233,7 @@ func TestParenthesizeVoid(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "void (a + b);")
 }
@@ -1259,7 +1258,7 @@ func TestParenthesizeTypeOf(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "typeof (a + b);")
 }
@@ -1284,7 +1283,7 @@ func TestParenthesizeAwait(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "await (a + b);")
 }
@@ -1407,7 +1406,7 @@ func TestParenthesizeBinary(t *testing.T) {
 					),
 				},
 			))
-			ast.SetParentInChildren(file)
+
 			parsetestutil.MarkSyntheticRecursive(file)
 			emittestutil.CheckEmit(t, nil, file.AsSourceFile(), rec.output+";")
 		})
@@ -1438,7 +1437,7 @@ func TestParenthesizeConditional1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a, b) ? c : d;")
 }
@@ -1467,7 +1466,7 @@ func TestParenthesizeConditional2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a = b) ? c : d;")
 }
@@ -1500,7 +1499,7 @@ func TestParenthesizeConditional3(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(() => { }) ? a : b;")
 }
@@ -1523,7 +1522,7 @@ func TestParenthesizeConditional4(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(yield) ? a : b;")
 }
@@ -1552,7 +1551,7 @@ func TestParenthesizeConditional5(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "a ? (b, c) : d;")
 }
@@ -1581,7 +1580,7 @@ func TestParenthesizeConditional6(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "a ? b : (c, d);")
 }
@@ -1607,7 +1606,7 @@ func TestParenthesizeYield1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "yield (a, b);")
 }
@@ -1643,7 +1642,7 @@ func TestParenthesizeSpreadElement1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "[...(a, b)];")
 }
@@ -1678,7 +1677,7 @@ func TestParenthesizeSpreadElement2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "a(...(b, c));")
 }
@@ -1711,7 +1710,7 @@ func TestParenthesizeSpreadElement3(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "new a(...(b, c));")
 }
@@ -1744,7 +1743,7 @@ func TestParenthesizeExpressionWithTypeArguments(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a, b)<c>;")
 }
@@ -1773,7 +1772,7 @@ func TestParenthesizeAsExpression(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a, b) as c;")
 }
@@ -1802,7 +1801,7 @@ func TestParenthesizeSatisfiesExpression(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a, b) satisfies c;")
 }
@@ -1828,7 +1827,7 @@ func TestParenthesizeNonNullExpression(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(a, b)!;")
 }
@@ -1849,7 +1848,7 @@ func TestParenthesizeExpressionStatement1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "({});")
 }
@@ -1878,7 +1877,7 @@ func TestParenthesizeExpressionStatement2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(function () { });")
 }
@@ -1902,7 +1901,7 @@ func TestParenthesizeExpressionStatement3(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "(class {\n});")
 }
@@ -1930,7 +1929,7 @@ func TestParenthesizeExpressionDefault1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "export default (class {\n});")
 }
@@ -1965,7 +1964,7 @@ func TestParenthesizeExpressionDefault2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "export default (function () { });")
 }
@@ -1991,7 +1990,7 @@ func TestParenthesizeExpressionDefault3(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "export default (a, b);")
 }
@@ -2020,7 +2019,7 @@ func TestParenthesizeArrayType(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = (a | b)[];")
 }
@@ -2055,7 +2054,7 @@ func TestParenthesizeOptionalType(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = [\n    (a | b)?\n];")
 }
@@ -2088,7 +2087,7 @@ func TestParenthesizeUnionType1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = a | (() => b);")
 }
@@ -2122,7 +2121,7 @@ func TestParenthesizeUnionType2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = (infer a extends b) | c;")
 }
@@ -2156,7 +2155,7 @@ func TestParenthesizeIntersectionType(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = a & (b | c);")
 }
@@ -2186,7 +2185,7 @@ func TestParenthesizeReadonlyTypeOperator1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = readonly (a | b);")
 }
@@ -2212,7 +2211,7 @@ func TestParenthesizeReadonlyTypeOperator2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = readonly (keyof a);")
 }
@@ -2242,7 +2241,7 @@ func TestParenthesizeKeyofTypeOperator(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = keyof (a | b);")
 }
@@ -2272,7 +2271,7 @@ func TestParenthesizeIndexedAccessType(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = (a | b)[c];")
 }
@@ -2303,7 +2302,7 @@ func TestParenthesizeConditionalType1(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = (() => a) extends b ? c : d;")
 }
@@ -2333,7 +2332,7 @@ func TestParenthesizeConditionalType2(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = a extends (b extends c ? d : e) ? f : g;")
 }
@@ -2371,7 +2370,7 @@ func TestParenthesizeConditionalType3(t *testing.T) {
 			),
 		},
 	))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = a extends () => (infer b extends c) ? d : e;")
 }
@@ -2414,7 +2413,7 @@ func TestParenthesizeConditionalType4(t *testing.T) {
 			),
 		),
 	}))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, nil, file.AsSourceFile(), "type _ = a extends () => (infer b extends c) | d ? e : f;")
 }
@@ -2446,7 +2445,7 @@ func TestNameGeneration(t *testing.T) {
 			}), true),
 		),
 	}))
-	ast.SetParentInChildren(file)
+
 	parsetestutil.MarkSyntheticRecursive(file)
 	emittestutil.CheckEmit(t, ec, file.AsSourceFile(), "var _a;\nfunction f() {\n    var _a;\n}")
 }