@jakebailey 2025-06-20
You know, I think I see why the parent setting is so slow. My fix in #403 reuses functions to avoid the closure allocation (since ForEachChild escapes its args), but since it's recursive, it can _never_ actually reuse them during a single run, only later in a completely separate call chain. There might be a way to speed this up and avoid the perf hit.